PSY-961: create-from-entity drawer pre-fill (PSY-893 D4)
Implements D4 via an app-level Create drawer (Option B — the in-place,
research-backed choice over query-param navigation). The popover's CTA
becomes "Create new collection with {entity}" / empty-state "Create with
{entity}" and opens the drawer pre-filled with the current entity as item
1 — closing the "I'm reading about this artist, start a collection around
them" loop without leaving the page.

- Extract CreateCollectionForm out of CollectionList into its own file,
  with an `initialStagedItems` prop that seeds the staged list.
- New CreateCollectionDrawer: a CreateCollectionDrawerProvider (mounted at
  the root via Providers) exposing useCreateCollectionDrawer().openCreateDrawer().
  The form is lazy-loaded via dynamic(ssr:false) so AddItemsPicker + its
  extraction/search deps stay out of the root chunk (chunk hygiene).
- Migrate CollectionList's local Sheet onto the shared drawer (one drawer).
- AddToCollectionButton CTAs call openCreateDrawer with the entity pre-filled.

Tests: CreateCollectionForm (render defaults, cover field, pre-fill, submit
→ create + bulk-add), CreateCollectionDrawer (throw-without-provider,
open-with-prefill), AddToCollectionButton (both CTAs pre-fill the entity),
CollectionList (button opens the drawer; form tests moved out), providers
(drawer provider mounts). 940/940 in scope.
